【bug】老是执行中断

+++ cublas_inv_matrix.cuh
@@ -2,6 +2,6 @@
 #define CUBLAS_INV_MATRIX_H
 
 
-int my_gauss_inverse_gpu_by_cublas(double** d_in, int size, double** d_out, int my_np);
+int my_gauss_inverse_gpu_by_cublas(float** d_in, int size, float** d_out, int my_np);
 
 #endif // !CUBLAS_INV_MATRIX_H
